refactor: move common refund logic into internal
