+++ e-token/src/processor/internal/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod ephemeral_ata;
 pub(crate) mod group_receipt_accounts;
 pub(crate) mod lamports_pda;
 pub(crate) mod queue_authorized_action;
+pub(crate) mod refund;
 pub(crate) mod shuttle_delegation;
 pub(crate) mod token_vault;
 pub(crate) mod transfer_queue_refill;
